api/requirements.txt
flask
google-auth
requests
python-dotenv
pymongo